+++ src/app/core/models/dto/credential-offer-response.ts
@@ -1,4 +1,5 @@
 export interface CredentialOfferResponse {
   credential_offer_uri: string;
   c_transaction_code: string;
+  c_transaction_code_expires_in: number;
 }