+++ src/app/features/credential-detail/credential-detail.component.ts
@@ -1,7 +1,7 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslatePipe, TranslateService } from '@ngx-translate/core';
-import { from, Observable, switchMap, take, tap } from 'rxjs';
+import { EMPTY, from, Observable, switchMap, take, tap } from 'rxjs';
 import { CredentialProcedureService } from 'src/app/core/services/credential-procedure.service';
 import { LEARCredentialEmployeeJwtPayload } from "../../core/models/entity/lear-credential-employee.entity";
 import { LearCredentialEmployeeDataDetail } from "../../core/models/dto/lear-credential-employee-data-detail.dto";
@@ -50,7 +50,9 @@ export class CredentialDetailComponent implements OnInit {
   }
 
   public loadCredentialDetail(procedureId: string): void {
-    this.credentialProcedureService.getCredentialProcedureById(procedureId).subscribe({
+    this.credentialProcedureService.getCredentialProcedureById(procedureId)
+    .pipe(take(1))
+    .subscribe({
       next: (credentials: LearCredentialEmployeeDataDetail) => {
         this.credential = credentials['credential'];
         this.credentialStatus = credentials['credential_status'];
@@ -59,36 +61,51 @@ export class CredentialDetailComponent implements OnInit {
         console.error('Error fetching credential details', error);
       }
     });
-   
   }
 
-  public sendReminder(): void {
-    const credentialId = this.credentialId;
-    if (!credentialId){
-      console.error('No credential id.');
-      return;
-    }
+  public openSendReminderDialog(): void {
 
     const dialogData: DialogData = {
       title: this.translate.instant("credentialDetail.sendReminderConfirm.title"),
       message: this.translate.instant("credentialDetail.sendReminderConfirm.message"),
       confirmationType: 'async',
-      status: 'default',
-      loadingData: undefined
+      status: 'default'
     };
 
-    const sendReminderAfterConfirm = (): Observable<any> => {
-      return this.credentialProcedureService.sendReminder(credentialId)
-        .pipe(
-          switchMap(()  => 
-            from(this.router.navigate(['/organization/credentials'])).pipe(
-              tap(() => location.reload())
-            )
-          ));
+    const sendReminderAfterConfirm = (): Observable<boolean> => {
+      return this.sendReminder();
     }
     
     this.dialog.openDialogWithCallback(dialogData, sendReminderAfterConfirm);
 
+  }
+
+  public sendReminder(): Observable<boolean>{
+    const credentialId = this.credentialId;
+    if (!credentialId){
+      console.error('No credential id.');
+      return EMPTY;
+    }
+
+    return this.credentialProcedureService.sendReminder(credentialId)
+    // open success dialog and navigate to credentials
+    .pipe(
+      switchMap(() => {
+            const dialogData: DialogData = {
+              title: this.translate.instant("credentialDetail.sendReminderSuccess.title"),
+              message: this.translate.instant("credentialDetail.sendReminderSuccess.message"),
+              confirmationType: 'none',
+              status: 'default'
+            };
+
+            const dialogRef = this.dialog.openDialog(dialogData);
+            return dialogRef.afterClosed();
+      }),
+      switchMap(()  => 
+        from(this.router.navigate(['/organization/credentials']))
+      ),
+      tap(() => location.reload())
+    );
   }
 
 }