+++ src/app/features/credential-offer/credential-offer-stepper/credential-offer-stepper.component.ts
@@ -2,13 +2,13 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
 import { CredentialOfferComponent } from '../credential-offer-steps/credential-offer/credential-offer.component';
 import { CredentialOfferOnboardingComponent } from '../credential-offer-steps/credential-offer-onboarding/credential-offer-onboarding.component';
-import { Component, computed, DestroyRef, effect, inject, OnInit, Signal } from '@angular/core';
+import { Component, computed, DestroyRef, effect, inject, OnInit, Signal, TemplateRef, ViewChild, ViewContainerRef } from '@angular/core';
 import { MatStepperModule } from '@angular/material/stepper';
 import { MatButtonModule } from '@angular/material/button';
 import { NavbarComponent } from 'src/app/shared/components/navbar/navbar.component';
 import { QRCodeModule } from 'angularx-qrcode';
 import { MatIcon } from '@angular/material/icon';
-import { catchError, filter, map, merge, Observable, of, scan, startWith, Subject, switchMap, throwError } from 'rxjs';
+import { catchError, EMPTY, filter, interval, map, merge, Observable, of, scan, shareReplay, startWith, Subject, switchMap, take, tap, throwError, timer } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
 import { DialogWrapperService } from 'src/app/shared/components/dialog/dialog-wrapper/dialog-wrapper.service';
 import { CredentialProcedureService } from 'src/app/core/services/credential-procedure.service';
@@ -16,7 +16,9 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { BreakpointObserver } from '@angular/cdk/layout';
 import { STEPPER_GLOBAL_OPTIONS } from '@angular/cdk/stepper';
 import { CredentialOfferResponse } from 'src/app/core/models/dto/credential-offer-response';
-import { TranslateService } from '@ngx-translate/core';
+import { TranslatePipe, TranslateService } from '@ngx-translate/core';
+import { AsyncPipe } from '@angular/common';
+import { TemplatePortal } from '@angular/cdk/portal';
 
 export type StepperIndex = 0 | 1;
 export type CredentialOfferStep = 'onboarding' | 'offer';
@@ -24,22 +26,45 @@ export interface CredentialOfferParams {
   credential_offer_uri: string|undefined,
   transaction_code: string|undefined,
   c_transaction_code: string|undefined,
+  c_transaction_code_expires_in: number|undefined //expected in seconds
 }
 export interface CredentialOfferParamsState extends CredentialOfferParams {
-  loading: boolean
+  loading: boolean,
+  error: boolean
 }
 
 export const undefinedCredentialOfferParamsState: CredentialOfferParamsState = { 
     credential_offer_uri: undefined,
     transaction_code: undefined,
     c_transaction_code: undefined,
-    loading: false
+    c_transaction_code_expires_in: undefined,
+    loading: false,
+    error: false
   };
+
+  //REFRESH-SESSION RELATED CONSTANTS
+  type StartOrEnd = 'START' | 'END';
+  //time before refresh offer popup is shown
+  export const defaultTotalAvailableTimeInMs = 60 * 1000 * 9; // 9min in miliseconds
+  //countdown for the refresh offer popup; when it comes to 0, redirects to home
+  export const popupTimeInSeconds = 60; //in seconds; should always be 60s
+  export const loadingBufferTimeInMs = 25 * 1000; //margin to compensate for loading time 
 
 @Component({
   selector: 'app-credential-offer-stepper',
   standalone: true,
-  imports: [CredentialOfferComponent, CredentialOfferOnboardingComponent, MatButtonModule, MatIcon, MatProgressSpinnerModule, MatStepperModule, NavbarComponent, QRCodeModule],
+  imports: [
+    AsyncPipe,
+    CredentialOfferComponent, 
+    CredentialOfferOnboardingComponent,
+    MatButtonModule, 
+    MatIcon, 
+    MatProgressSpinnerModule, 
+    MatStepperModule, 
+    NavbarComponent, 
+    QRCodeModule,
+    TranslatePipe
+    ],
   providers: [{
     provide: STEPPER_GLOBAL_OPTIONS, useValue: {displayDefaultIndicatorType: false}
 }],
@@ -47,6 +72,7 @@ export const undefinedCredentialOfferParamsState: CredentialOfferParamsState = {
   styleUrl: './credential-offer-stepper.component.scss',
 })
 export class CredentialOfferStepperComponent implements OnInit{
+  @ViewChild('popupCountdown') popupCountdown!: TemplateRef<any>;
   private readonly breakpointObserver = inject(BreakpointObserver);
   private readonly credentialProcedureService = inject(CredentialProcedureService);
   private readonly destroyRef = inject(DestroyRef);
@@ -55,18 +81,19 @@ export class CredentialOfferStepperComponent implements OnInit{
   private readonly router = inject(Router);
   private readonly translate = inject(TranslateService);
 
-  //Source actions
+
+  //Subjects
   public getInitUrlParams$$ = new Subject<void>();
   public updateIndex$$ = new Subject<StepperIndex>();
   public loadCredentialOfferOnRefreshClick$$ = new Subject<void>();
   
-  //Effects
-  public updateUrlParamsOnOfferChange = effect(()=> {
+  //Offer effects
+  public updateUrlParamsOnOfferChangeEffect = effect(()=> {
     const offer = this.offerParams$();
     this.updateUrlParams(offer);
   });
 
-  //States
+  //Derived streams
   public currentIndex$ = toSignal<StepperIndex>(this.updateIndex$$.pipe(startWith(0 as StepperIndex)));
   public currentStep$ = computed<CredentialOfferStep>(() => {
     return this.currentIndex$() === 0 ? 'onboarding' : 'offer';
@@ -82,17 +109,20 @@ export class CredentialOfferStepperComponent implements OnInit{
     this.updateIndex$$.pipe(filter(() => this.currentIndex$() === 1)))
   .pipe(
     switchMap(()=>this.getCredentialOffer().pipe(
-      map(params => ({...params, loading: false })),
+      map(params => ({...params, loading: false, error: false })),
       catchError(error => { 
         console.error(error);
         return of({ 
-          loading: false
+          loading: false,
+          error: true
         })
       }),
       startWith({
-        loading: true
+        loading: true,
+        error: false
       })
     )),
+    shareReplay()
   );
 
   public offerParams$: Signal<CredentialOfferParamsState> = toSignal(
@@ -106,7 +136,7 @@ export class CredentialOfferStepperComponent implements OnInit{
         undefinedCredentialOfferParamsState
       )
   ), 
-  { initialValue: undefinedCredentialOfferParamsState }); //this is needed because the seed value in scan is not emitted
+  { initialValue: undefinedCredentialOfferParamsState }); //this is needed because the seed value in scan() is not emitted
 
   public stepperOrientation$ = toSignal(this.breakpointObserver
     .observe('(min-width: 800px)')
@@ -115,9 +145,101 @@ export class CredentialOfferStepperComponent implements OnInit{
     initialValue: 'horizontal'
   });
 
-public ngOnInit(): void {
-  this.getInitUrlParams$$.next();
-}
+  //EXPIRATION-RELATED STREAMS
+  //when this startOrEndFirstCountdown$ emits 'START', refresh popup is closed and the endsession countdown starts
+  //when it emits 'END', the popup is closed and the countdown is interrupted
+  //if user clicks to refresh session, new offer params are fetched, which makes restart the previous flow
+  //if there is an error when refreshing session, user is redirected to home
+  //if user clicks to leave session, user is redirected to home
+  private readonly startOrEndFirstCountdown$: Observable<StartOrEnd> = this.fetchedCredentialOffer$
+  .pipe(
+    filter(offerState => (offerState.loading === false) && (offerState.error === false)),
+    switchMap(offerParams => {
+      const totalAvailableTimeFromBackendInSeconds = offerParams.c_transaction_code_expires_in;
+      let totalAvailableTimeInMs: number;
+      if(!totalAvailableTimeFromBackendInSeconds){
+        console.error('Offer expiration time not received from API; using default: ' + defaultTotalAvailableTimeInMs + ' - ' + loadingBufferTimeInMs);
+        totalAvailableTimeInMs = defaultTotalAvailableTimeInMs;
+      }else{
+        totalAvailableTimeInMs = (totalAvailableTimeFromBackendInSeconds * 1000);
+      }
+      return timer(totalAvailableTimeInMs - (popupTimeInSeconds * 1000) - loadingBufferTimeInMs).pipe(
+        map(() => 'END' as StartOrEnd),
+        startWith('START' as StartOrEnd)
+      )
+    }),
+    shareReplay()
+  );
+
+  private readonly openRefreshPopupEffect = this.startOrEndFirstCountdown$.pipe(
+    filter( val => val === 'END'),
+    tap(() => {
+      this.dialog.openDialogWithCallback({
+        title: this.translate.instant('credentialOffer.expired-title'), 
+        message: '',
+        template: new TemplatePortal(this.popupCountdown, {} as ViewContainerRef),
+        confirmationType: 'async',
+        status: 'default', //error?
+        confirmationLabel: 'Refresh',
+        cancelLabel: 'Leave'
+      }, 
+      //after confirmation callback
+      () => {
+        this.onRefreshCredentialClick(); 
+        return EMPTY;
+      }, 
+      //after cancel callback
+      () => {
+        this.redirectToHome();
+        return EMPTY;
+      },
+      //avoid closing popup clicking on overlay (force to click leave or refresh button)
+      'DISABLE_CLOSE'
+  )})
+  );
+
+  private readonly closeRefreshPopupEffect = this.startOrEndFirstCountdown$.pipe(
+    filter( val => val === 'START'),
+    tap(() => {
+      this.dialog['dialog'].closeAll();
+    })
+  );
+
+  public readonly endSessionCountdown$ = this.startOrEndFirstCountdown$.pipe(
+    switchMap(startOrEnd => {
+      if(startOrEnd === 'END'){
+        return interval(1000).pipe(
+          take(popupTimeInSeconds + 1),
+          map(consumedSeconds => popupTimeInSeconds - consumedSeconds)
+        )
+      }else{
+        return of(popupTimeInSeconds); //reset time
+      }
+    }), 
+    shareReplay()
+  )
+
+  private readonly navigateHomeAfterEndSessionEffect = this.endSessionCountdown$.pipe(
+    filter(time => time === 0),
+    tap(()=>{
+      this.redirectToHome();
+      this.dialog.openErrorInfoDialog(this.translate.instant("error.credentialOffer.expired"));
+    })
+  );
+
+  //END EXPIRATION TIME STREAMS
+
+  public ngOnInit(): void {
+    this.getInitUrlParams$$.next();
+    
+    merge(
+      this.openRefreshPopupEffect,
+      this.closeRefreshPopupEffect,
+      this.navigateHomeAfterEndSessionEffect
+    )
+    .pipe(takeUntilDestroyed(this.destroyRef))
+    .subscribe();
+  }
 
   public onSelectedStepChange(index:number){
     if(((index !== 0) && (index !== 1))){
@@ -134,12 +256,22 @@ public ngOnInit(): void {
 
     if(!transactionCodeParam){
       console.error("Client error: Missing transaction code in the URL. Can't get credential offer.");
+      
+      this.translate.get("error.credentialOffer.invalid-url")
+      .pipe(take(1))
+      .subscribe((message:string)=>{
+        this.dialog.openErrorInfoDialog(message);
+        this.redirectToHome();
+        }
+      );
     }
     const updatedParams: CredentialOfferParamsState = {
       credential_offer_uri: undefined,
       transaction_code: transactionCodeParam,
       c_transaction_code: cCodeParam,
-      loading: false
+      c_transaction_code_expires_in: undefined,
+      loading: false,
+      error: false
     };
     return updatedParams;
   }
@@ -154,7 +286,7 @@ public ngOnInit(): void {
       params = this.getCredentialOfferByTransactionCode(offer.transaction_code);
     }else{
       this.redirectToHome();
-      console.log("Client error: Transaction code not found. Can't get credential offer");
+      console.error("Client error: Transaction code not found. Can't get credential offer");
       const message = this.translate.instant("error.credentialOffer.invalid-url");
       this.dialog.openErrorInfoDialog(message);
     }
@@ -163,7 +295,7 @@ public ngOnInit(): void {
 
   public getCredentialOfferByTransactionCode(transactionCode:string): Observable<CredentialOfferResponse> {
     if(!transactionCode){
-      console.log("No transaction code was found, can't refresh QR.");
+      console.error("No transaction code was found, can't refresh QR.");
       const message = this.translate.instant("error.credentialOffer.invalid-url");
       this.dialog.openErrorInfoDialog(message);
       this.redirectToHome();
@@ -191,7 +323,7 @@ public ngOnInit(): void {
 
   public getCredentialOfferByCTransactionCode(cCode:string): Observable<CredentialOfferResponse> {
     if (!cCode) {
-      console.log("No c-transaction code was found, can't refresh QR.");
+      console.error("No c-transaction code was found, can't refresh QR.");
       const message = this.translate.instant("error.credentialOffer.invalid-url");
       this.dialog.openErrorInfoDialog(message);
       this.redirectToHome();