+++ src/app/shared/components/dialog/dialog.component.ts
@@ -15,17 +15,23 @@ export interface LoadingData { title:string, message:string; template?:Component
 export interface DialogData { 
   title: string; 
   message: string; 
-  confirmationType: DialogConfirmationType;
   status: DialogStatus;
-  loadingData: LoadingData | undefined;
+  confirmationType: DialogConfirmationType;
+  template?: ComponentPortal<any>|TemplatePortal|DomPortal;
+  loadingData?: LoadingData;
+  confirmationLabel?: string;
+  cancelLabel?: string;
 }
 export interface DialogDefaultContent {
   data: DialogData;
 }
 
+//Confirmation type:none - no confirmation needed, so only one button is displayed, which closes the dialog
+//
 @Component({
     selector: 'app-dialog',
     templateUrl: './dialog.component.html',
+    //for styles, look _dialogscss in global styles
     standalone: true,
     imports: [
         AsyncPipe,
@@ -55,23 +61,41 @@ export class DialogComponent {
     this.dialogRef.addPanelClass('dialog-custom');
     this.updateStatus();
   }
-
-  public updateStatus(){
+  public updateStatus(): void{
     const previousStatus = this.currentStatus;
-    if(previousStatus){
-      this.dialogRef.removePanelClass(`dialog-${previousStatus}`);
-    }
     this.currentStatus = this.data.status;
-    this.dialogRef.addPanelClass(`dialog-${this.currentStatus}`);
+    this.updateStatusPanelClass(previousStatus);
+    this.updateStatusColor();
+  }
+  
+  //ideally this should be done with reactive state management
+  public updateStatusColor(): void{
+    this.statusColor = this.currentStatus === 'error' ? 'warn' : 'primary';
+  }
+
+  public updateStatusPanelClass(previousStatus: DialogStatus | undefined): void{
+    if(previousStatus !== this.currentStatus){
+      this.dialogRef.removePanelClass(`dialog-${previousStatus}`);
+      this.dialogRef.addPanelClass(`dialog-${this.currentStatus}`);
+    }
   }
 
   public updateData(data: Partial<DialogData>){
-    this.data = { ...this.data, ...data };
+    const resetDefaultOptionalData:Partial<DialogData> = {
+      template: undefined,
+      confirmationLabel: undefined,
+      cancelLabel: undefined
+    }
+    
+    this.data = { 
+      ...this.data, 
+      ...resetDefaultOptionalData,
+      ...data };
     this.updateStatus();
-
   }
 
-  public afterConfirmSubj(): Observable<boolean>{
+  // this is used by the dialog wrapper service, which subscribes to this subject to call the callback after the this emits 
+  public afterConfirm$(): Observable<boolean>{
     return this.confirmSubj$.asObservable();
   }