+++ src/app/shared/components/navbar/navbar.component.ts
@@ -3,6 +3,7 @@ import { Router, RouterLink } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import { AuthService } from 'src/app/core/services/auth.service';
 import { MatIcon } from '@angular/material/icon';
+import { take } from 'rxjs';
 
 @Component({
     selector: 'app-navbar',
@@ -21,27 +22,31 @@ export class NavbarComponent implements OnInit {
   private readonly router = inject(Router);
 
   public ngOnInit() {
-    this.translate.addLangs(['en', 'es', 'ca']);
-    this.translate.setDefaultLang('en');
-    this.selectedLanguage = this.translate.getDefaultLang();
-    this.authService.getMandator().subscribe(mandator => {
-      if (mandator) {
-        this.organization = mandator.organization
-      }
-    })
-    this.authService.getName().subscribe(name => {
-      if (name) {
-        this.userName = name;
-      }
-    });
+    this.authService.getMandator()
+      .pipe(take(1))
+      .subscribe(mandator => {
+        if (mandator) {
+          this.organization = mandator.organization
+        }
+      })
+    this.authService.getName()
+      .pipe(take(1))
+      .subscribe(name => {
+        if (name) {
+          this.userName = name;
+        }
+      });
   }
 
   public logout() {
-    this.authService.logout().subscribe(() => {
-      this.router.navigate(['/home'], {});
-    });
+    this.authService.logout()
+      .pipe(take(1))
+      .subscribe(() => {
+        this.router.navigate(['/home'], {});
+      });
   }
 
+  //currently not used
   public changeLanguage(languageCode: string): void {
     this.translate.use(languageCode);
     this.selectedLanguage = languageCode;